Huge Code Cleanup

+++ Cirilla/CommandLogger.cs
@@ -1,10 +1,10 @@
-﻿using Cirilla.Services.Pastebin;
-using Discord;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using Cirilla.Services.Pastebin;
+using Discord;
 
 namespace Cirilla {
     public static class CommandLogger {
@@ -16,17 +16,14 @@ namespace Cirilla {
                 lock (Lock) {
                     try {
                         string logfile = Path.Combine(Information.Directory, "commandlog.txt");
-                        if (!File.Exists(logfile)) {
-                            File.Create(logfile).Dispose();
-                        }
+                        if (!File.Exists(logfile)) File.Create(logfile).Dispose();
 
                         // > than 10 MB (default) -> clear log
-                        if (Information.Config != null && new FileInfo(logfile).Length > Information.MaxLogSize) {
+                        if (Information.Config != null && new FileInfo(logfile).Length > Information.MaxLogSize)
                             File.WriteAllBytes(logfile, new byte[0]);
-                        }
 
                         File.AppendAllLines(logfile,
-                            new List<string> { $"[{DateTime.Now:HH:mm:ss}] [{username}@{guildname}] \"{command}\"" });
+                            new List<string> {$"[{DateTime.Now:HH:mm:ss}] [{username}@{guildname}] \"{command}\""});
                     } catch (Exception ex) {
                         Console.ForegroundColor = ConsoleColor.Red;
                         Console.WriteLine($"Could not save Log to Command Log File! ({ex.Message})");
@@ -55,9 +52,7 @@ namespace Cirilla {
             lock (Lock) {
                 try {
                     string logfile = Path.Combine(Information.Directory, "commandlog.txt");
-                    if (!File.Exists(logfile)) {
-                        File.Create(logfile).Dispose();
-                    }
+                    if (!File.Exists(logfile)) File.Create(logfile).Dispose();
                     long size = new FileInfo(logfile).Length / 1024;
                     File.WriteAllBytes(logfile, new byte[0]);
                     return size;
